Use probability curve to weight edges
Because the scoring system is probability based, and we have a spike of
uncertainty around 0.25, use the same curve as we do for target
selection to weight the edges of the graph. This should ensure that the
path should go from nodes with high confidence scores to nodes with low
confidence scores.

It might be even better to use a different curve that inverts the
parabola for the 'low' side, but this should at least be an improvement.