Make blockchain:ledger_at consistent about context creation
Prior to this change, the result of ledger_at might be the current
ledger with the existing context. This is problematic for cases where
the passed in ledger might have an existing context and code assumed
that the behaviour of ledger_at was always to give you the ledger as it
should have appeared at that block height.