Improve block signature checking
Prior to this change, we were tolerating signatures from unknown signers
as well as signatures that were invalid. This change requires all the
attached signatures to be valid, to be from one of the consensus members
and also requires at least 2f+1 valid signatures, and no more than 3f+1
signatures.

This change exposed some issues with the unit tests, so they had to be
corrected to supply only valid block signatures.