Implement 'assume valid' sync speed boost (#217)
When syncing the chain, if an 'assumed valid' block has been configured, don't validate blocks that come in until a complete chain from the tip of the main chain to the assumed valid block has been accumulated in a temporary block column family. At that point, if a valid hash chain of blocks can be constructed from the main chain's tip to the assumed valid block, absorb all the blocks without verifying them.

This is based on the design of assumed valid blocks from bitcoin, see https://github.com/bitcoin/bitcoin/pull/9484 for more details.