Fixes to assume valid to avoid OOM kills and unnecessary syncs
Prior to this change, 2 bugs were observed with assume_valid:

First, when accumulating a lot of blocks in memory to be absorbed, OOM
events were occuring. This has been fixed by only accmulating the block
hashes in memory and then only keeping the block in memory long enough
to absorb it to the ledger and save it to the block db.

Second, when the node got the designated assumed valid block, it would
not save that block to the assume valid block scratch space but instead
only retain it in memory. This meant that if the node crashed during the
assume-valid absorb, it would have to wait for another sync event to
happen that contained the assume-valid block before it would resume
absorbing the assume-valid blocks. This has been rectified by saving the
final assume-valid block to the scratch space.