Make assume-valid absorbtion always occur in a background process
Prior to this change, if the blockchain was interrupted mid async-valid
absorb, on startup the main blockchain init would be blocked absorbing
blocks. This is not necessary and it impedes the blockchain doing other
things while assume-valid processes in the background.

This change simply runs the assume-valid absorb in a background process,
much like when it's triggered by a blockchain sync.