Intentionally break compatability with the old sync handler protocol
This change has 2 purposes; to prevent older hotspots with buggy
firmware from syncing the blockchain until they upgrade and to
improve the network characteristics of syncing.

The improved characteristcs of syncing are to send blocks in smaller
batches of a configurable size, up to a configurable limit of batches
sent overall. This prevents too much state buildup in a libp2p muxer and
possibly dropping the payload without receiving any blocks. Additionally
this allows syncs that are not useful or are incorrect to fail much
earlier than the old protocol.