Add transaction bundling
Transactions that need a relative order can now be bundled together
before being submitted to the chain. This ensures that the transactions
don't get separated and the dependent transactions won't be rejected
because the transactions they depend on have not been absorbed.

Transaction bundles must consist of all valid and absorbable
transactions and will appear on chain still bundled. Bundles appear at
the end of the transaction list.