Cache ledger snapshots and use them for ledger_at, when available
Since we already create ledger snapshots on block absorb, cache them in
an ETS table so that if we need to compute a ledger at that height
later, we have it cached. Snapshots older than the lagging ledger will
be deleted from the cache and be garbage collected as normal. This
provides a very sigificant speed increase when syncing blocks or
when the consensus group is validating proposed transactions.

In addition remove a log message from poc receipt validation that was
accidentally merged.