validation timings
we've been having some issues telling what parts of the validation are
the most expensive.  this patch adds some opt-int debug logging for
validation times, and drills down on the receipt validation, which is
a known challenge.  this code is due to be removed when we have some
better overall design, but is meant as a stopgap to make it easier to
debug and observe in hotspots in the field