Implement 'context' snapshots to improve startyp performance
While the ledger snapshot cache is very fast and useful, when a node
starts up it doesn't have the cache available to it and so it must
laboriously compute the ledger for all the snapshot cache misses. Until
the node has processed ~20 blocks, this is extremely slow.

This patch introduces a second kind of snapshot in the cache, a
'context' snapshot, which is a snapshot of a ledger context as created
by 'ledger_at' by replaying transactions on top of the lagging ledger.
This is not as fast as a proper snapshot but it can still provide
signifant performance improvements until the node has processed enough
blocks to have real snapshots available.