Switch witness garbage collection to on read rather than on assert
Prior to this change, gateway witnesses were garbage collected when
their nonce changed as part of a location assertion. This involved a
full scan of all the ledger gateways, which is at least O(N).

This change attempts to drastically reduce the scope by checking the
nonces of witnesses against the actual gateway record's nonce when the
gateway record is loaded from rocksdb. Since gateways have a much
smaller witness list than the total number of gateways this bounds the
compute complexity much lower.