adjust gossip/sync strategy
we've been having some issues with hotpots falling out of sync due to
some prior bugfixing.  there is a 5 minute cooldown on starting a new
sync, added when we were having some clock skew issues that were
causing spots to never manage to sync.  unfortunately, it means that
if a spot misses timely gossip, which happens often, it can't sync up
for five minutes after the last block.  often, when the spot does
sync, it then misses gossip again and has to wait another five minutes
to have another attempt to get back onto the gossip train.

this patch tries, when we're not obviously clock skewed, to use the
block time rather than the last-added block time.  this should lead
to quicker sync cycles and make it more likely that the spot gets back
into sync and stays there.