Properly sequence dependent transactions in the transaction manager
When delivering a transaction to the consensus group, check if any other
transactions current in-progress are dependencies. If any are found,
submit the transaction only to members which have accepted all the
dependencies. This ensures that the dependent transactions appear in the
same buffer as their dependencies or they wait until all the
dependencies have cleared.