add gateway cache
Profiling has indicated that a huge amount of time is being spent
reading gateways off the disk or out of the cache and deserializing
them.  This PR provides a context-safe gateway cache that stores a
limited number of height indexed gateways in ETS so that they can be
fetched more cheaply.

Additionally there has been some work to make the validation fo PoC
packets cheaper, as that was the next thing that bubbled to the top of
the profile once deserialization had been largely taken care of.

Further work will need to be done in lessening the cost of determining
the current ledger height for the cache, which typically has to go all
the way to the disk.