Run ledger GC hooks in the absorb context
Prior to this change, the GC hooks ran in their own context, which could
possibly lead them to being missed when we took the snapshot for the
add_block event and saved the snapshot in the cache. This reworks the GC
hooks to run in the before commit hook of absorb_and_commit so they can
run in the same ledger context for all the other block changes.