Add memory limit when constructing snapshots & coalesce duplicates
Snapshot generation can use a lot of memory which can cause OOMs on
smaller systems, this change bounds how much heap can be used to
generate a snapshot (default 200mb, set to 0 to disable). Additionally
if the same snapshot is being generated twice at the same time, the
snapshot is only constructed once to bound the memory usage.