Attempt to rework the server side of the SC handler
The SC handler was too parallel and it lost track of offers and
corresponding packets. Additionally it was validating packets more
thoroughly than offers. This branch attempts to require an offer to be
followed up with a matching packet, if the offer is accepted. If the
client sends another offer, that second offer is queued pending the
first offer being processed and the corresponding packet being sent, if
purchased.