When closing a disputed state channel slash overcommit & normalize payouts
Prior to this change, gated behind rewards v4, when a state channel
closed in dispute state, the DCs were paid out 1:1, even if the channel
closed with a total above the amount of DC in the state channel.
Additionally, the overcommit would be shared between all the
participants of the state channel, but that incentivizes the router to
pack the state channel with affiliated hotspots to try to recover some
of the overcommit and payout. With v4 we fully slash the overcommit and
proportionally pay out the participants (if the state channel is
overpaid), after truncating any large participant payouts to the size of
the state channel.