Implement a DC to HNT scaling factor when paying rewards in 1:1 mode
Currently there exists an arbitrage opportunity when the value at which
you burned HNT into DC is lower than the value you spend it. This PR
adds a scaling factor (which was originally proposed in HIP 10 but later
removed) so that arbitrage amounts under some percentage can be
prevented. For example, if this value was set to 0.9 there would have to
be a greater than 10% spread between the value of HNT at burn time and
the value of HNT at spend time to be able to make a profit. Naturally
this does not close long-term arbitrage holes but it prevents
front-running the oracle price in all but the most extreme adjustment
periods.