Codify implicit sorting of oracle submissions before txn close & fix issue
The implicit txn ordering behaviour if a transaction did not have a
defined sort order can cause instability in the sort order when a new
transaction is added, because the undefined sort order was the same as
the last sort order. This could lead to the transaction without a
defined sort order sorting before the transaction with a defined sort
order (because of the other sort dimensions like nonces, etc).This means
that adding another transaction to the end of the sort order list would
cause the transaction with the implicit sort order to definitely sort
after the transaction it may have sorted before earlier, invalidating
the block.

This change establishes a definitive sort order for the oracle price
report txn to mirror the implicit sort order before the addition of the
hotspot transfer function. Additionally it makes the undefined sort
order one past the maximum to prevent this bug recurring and it add a
log message to the sync_handler to make this failure clearer.

This *may* expose other implicit sort ordering bugs, so a thorough
resync (or at least ordering check of txns in blocks) is recommended.