don't fetch a million blocks for their height
this PR both increases the pace of poc gc (trying to make them cheaper), and also fetches only the block hash for the relevant blocks from the disk, which should make the gc dramatically cheaper.

it also removes a large sync write from the state channel client which doesn't seem to be required.