Optimization for searching for poc_requests
In get_channels we were using `blockhain:head_block` to start looking for poc_requests for a given onion hash.

This works for normal setups but when a blockchain.db is available that has many more blocks than the current ledger height, a _huge_ perfomance hit is incurred by searching from a much later block back to the current one before going back in height to actually fine the poc_request within a few blocks.

I believe this change is  completely backwards compatible