Test passing
the trick was adding a little bit of time after the packet was sent to
give it a chance to be processed. There was a very common race condition
where a packet would be sent from the second gateway, and block would be
added before it was proccessed into the state channel, causing the
closetxn check to fail.

I messed with the timing and could go down to 20ms before things started
to be noticeably inconsistent.