Make partial snapshot loads very obvious and avoidable
When checking if we need to load a snapshot we look at the ledger
height. Because snapshot loads are no longer atomic, for memory limit
reasons, put the height into the ledger as the last step so partial
snapshot loads are never mistaken for completed ledger loads.

Additionally we remove a pointless attempt to save unloaded ledgers that
could lead to issues with non-atomic loads when the height is loaded
last.