Fix save_plausible_blocks to not return an error if some blocks are plausible
The old version of save_plausible_blocks considered an existing block as
an error and would not count it a a plausible block. This led to cases
where the call to check_plausible_blocks did not run (for example it was
interrupted by a node shutdown). This led to cases where all the batches
of received blocks were plausible but already stored, which leads to the
sync handler never calling check_saved_blocks and thus a sync halt.

This patch refines the block checking to simply avoid saving a plausible
block we already have, but not considering it an error, thus allowing
the sync handler to run check_saved_blocks.