If a snapshot listed in latest.json is interrupted when loading, fix it
Previous to this change if a latest.json snap was loading and it was
interrupted (OOM, unplug, power loss, etc) the node would get stuck
trying to sync with an incomplete chain because the code didn't check
the latest.json snap when initializing the chain, only once the chain
had started. This would lead to a blessed snapshot in the config file
being "older" than the chain, but the chain being incomplete as a result
of an interrupted snapshot load from a latest.json snapshot at a higher
height.