Don't invalidate h3->region lookups on var nonce change
Previously all the h3->region lookups would be invalidated when any
chain var was changed. This was very expensive when unrelated vars
changed.

This patch changes things so that the lookup is keyed by the hash of the
h3 region binaries, so as long as those do not change the cache should
not be invalidated. It also adds caching for the h3 region binary hashes
and for the h3 binaries themselves.