Honor blockchain_txn:block_delay() when using ledger_at
For some unknown reason, some nodes may have an older delayed ledger
than the prescribed 50 blocks, this was leading to block disagreement
when a transaction whose validity was bounded by ledger availability was
included in a block and the validity needed a ledger 51+ blocks back.

This patch forces ledger_at to only ever return 50 blocks back, even if
the lagging ledger is lagging further behind.