Use ledger checkpoint/snapshot at currentheight - block_delay for snaps
Prior to this the default way to take a snapshot was to use the delayed
ledger. However recent events has revealed delayed ledgers lagging too
far behind and delayed ledger corruption. In the time since the delayed
ledger was added the chain has switched to using rocksdb
snapshots/checkpoints which offer a more reliable and less vulnerable to
corruption source for snapshot data.