Add a chain var to isolate var txns to their own blockchain_vars
Due to several observed cache issues when a var txn appears in a block
with other transactions that may have been affected by the var changing,
this variable has been introduced to force var txns to appear in their
own block. This is done as a var so that it can only be activated once a
majority of the consensus group understands the var, so a split block
will not be produced.